Made coursehistories check for equality via ids rather than name. Aliases really need to be fixed now.

+++ app/templatetags/scorecard_tag.py
@@ -12,7 +12,7 @@ register = template.Library()
 class ScoreBox(object):
   def __init__(self, description, number):
     try:
-      self.number = "%.1f" % int(number)
+      self.number = "%.1f" % float(number)
     except:
       self.number = number
     self.description = description
